Pass model to widget. Use callbacks for midi i/o. SVG Buttons
Passing the model to the widget means each widget can use it for the
necessary parameters. It's simpler than passing parameter strings to
Section.add_class, and much less ugly. The widget constructors know
what to do.

For midi i/o, it doesn't really make sense to add a timer and buffer all
the midi params. Iterating over the changed widgets and tracking that
info seems like needless overhead, given that events still need to be
handled, and control signals incoming off of a controller are less likely
to be frequent. Generally, the ui needs to keep up with transmitting.

Now interface manager calls updated_cc or updated_nrpn on the appropriate
MidiControl object (based on which port the object is set to).

MidiControl then simply calls the widget.update function with the new
value.